+++ example/react-example/.ddotrc.ts
@@ -3,27 +3,6 @@ export default {
   plugins: [
     './test-plugin',
     [
-      'jenkins',
-      {
-        hostName: 'jenkins.hello.com',
-        pathPrefix: '/job/job',
-        prompt: ({ branch, jobName }) => [
-          {
-            type: 'list',
-            name: 'Branch',
-            messages: 'which branch to deploy',
-            choices: () => [branch, 'develop', 'master'],
-          },
-          {
-            type: 'list',
-            name: 'Env',
-            messages: 'which Env to deploy',
-            choices: () => ['develop', 'release'],
-          },
-        ],
-      },
-    ],
-    [
       'webpack',
       {
         proxy: {