+++ example/vue-example/extend.d.ts
@@ -1,0 +1,4 @@
+declare module '*.vue' {
+  const content: string;
+  export default content;
+}