+++ example/vue-example/src/index.ts
@@ -1,0 +1,7 @@
+import Vue from 'vue';
+import App from './pages/App.vue';
+new Vue({
+  render(h) {
+    return h(App);
+  },
+}).$mount('#root');