+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.6](https://github.com/Jetsly/ddot/compare/@ddot/ddot-cli@0.0.5...@ddot/ddot-cli@0.0.6) (2018-12-14)
+
+**Note:** Version bump only for package @ddot/ddot-cli
+
+
+
+
+
 ## [0.0.5](https://github.com/Jetsly/ddot/compare/@ddot/ddot-cli@0.0.4...@ddot/ddot-cli@0.0.5) (2018-12-13)