+++ packages/plugin-jenkins/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.3](https://github.com/Jetsly/ddot/compare/@ddot/ddot-plugin-jenkins@0.0.2...@ddot/ddot-plugin-jenkins@0.0.3) (2018-12-14)
+
+
+### Bug Fixes
+
+* **ddot-plugin-jenkins:** remove shelljs dep ([60f1179](https://github.com/Jetsly/ddot/commit/60f1179))
+
+
+
+
+
 ## 0.0.2 (2018-12-13)