+++ packages/plugin-webpack/CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.12](https://github.com/Jetsly/ddot/compare/@ddot/ddot-plugin-webpack@0.0.11...@ddot/ddot-plugin-webpack@0.0.12) (2018-12-14)
+
+
+### Features
+
+* **ddot-plugin-webpack:** auto load src/pages/document.ejs ([055644a](https://github.com/Jetsly/ddot/commit/055644a))
+
+
+
+
+
+## [0.0.11](https://github.com/Jetsly/ddot/compare/@ddot/ddot-plugin-webpack@0.0.10...@ddot/ddot-plugin-webpack@0.0.11) (2018-12-14)
+
+
+### Bug Fixes
+
+* **ddot-plugin-jenkins:** remove shelljs dep ([60f1179](https://github.com/Jetsly/ddot/commit/60f1179))
+* **ddot-plugin-webpack:** fix publicpatch prefix ([7995af2](https://github.com/Jetsly/ddot/commit/7995af2))
+
+
+
+
+
 ## [0.0.10](https://github.com/Jetsly/ddot/compare/@ddot/ddot-plugin-webpack@0.0.9...@ddot/ddot-plugin-webpack@0.0.10) (2018-12-13)